referrer-policy/generic/inheritance/iframe-inheritance-about-blank.html

<!doctype html>
<meta charset="utf-8">
<title>Referrer Policy: iframe src="about:blank"</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="referrer" content="origin">
<body>
<script>
const testClient = async_test("referrer=client");
const testUrl = async_test("referrer=URL");

window.addEventListener("message", msg => {
  if (msg.data.name === "client") {
    testClient.step_func_done(() => {
      // Because the URL of the Document of <iframe src="about:blank"> is
      // "about:blank", the stripped URL is no referrer:
      // https://w3c.github.io/webappsec-referrer-policy/#strip-url
      assert_equals(msg.data.referrer, undefined);
    })();
  } if (msg.data.name === "URL") {
    // <iframe src="about:blank"> doesn't inherit parent's referrer policy.
    // Note: Setting an explicit URL as referrer succeeds
    // because the same-origin check at
    // https://fetch.spec.whatwg.org/#dom-request
    // is done against <iframe>'s origin, which inherits the parent
    // Document's origin == location.orgin.
    testUrl.step_func_done(() => {
      assert_equals(msg.data.referrer, location.href + '/custom');
    })();
  }
});

const iframe = document.createElement("iframe");
document.body.appendChild(iframe);

const script2 = iframe.contentDocument.createElement('script');
script2.innerText = `
  fetch("${location.origin}/common/security-features/subresource/xhr.py?name=client")
    .then(r => r.json())
    .then(j => {
        top.postMessage({name: "client", referrer: j.headers.referer}, "*")
      }).catch(e => {
        top.postMessage({name: "client", referrer: "FAILURE"}, "*");
      });

  fetch("${location.origin}/common/security-features/subresource/xhr.py?name=URL",
      {referrer: "${location.href}/custom"})
    .then(r => r.json())
    .then(j => {
        top.postMessage({name: "URL", referrer: j.headers.referer}, "*")
      }).catch(e => {
        top.postMessage({name: "URL", referrer: "FAILURE"}, "*");
      });
`;
iframe.contentDocument.body.appendChild(script2);
</script>
